ahora si

+++ src/main/java/com/babas/models/Company.java
@@ -23,15 +23,15 @@ public class Company extends Babas {
     private String directionPrincipal;
     private String logo;
     private String slogan;
-    private Long idFact;
-    private Long idProduct;
+    private Integer idFact;
+    private Integer idProduct;
     private String token;
 
-    public Long getIdProduct() {
+    public Integer getIdProduct() {
         return idProduct;
     }
 
-    public void setIdProduct(Long idProduct) {
+    public void setIdProduct(Integer idProduct) {
         this.idProduct = idProduct;
     }
 
@@ -43,11 +43,11 @@ public class Company extends Babas {
         this.token = token;
     }
 
-    public Long getIdFact() {
+    public Integer getIdFact() {
         return idFact;
     }
 
-    public void setIdFact(Long idFact) {
+    public void setIdFact(Integer idFact) {
         this.idFact = idFact;
     }
